llamaindex/pyproject.toml
[project]
name = "aeris-llamaindex"
version = "0.1.0"
description = ""
authors = [{ name = "Erik Peterson", email = "e@eriklp.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "llama-index (>=0.12.14,<0.13.0)",
    "llama-index-readers-weather (>=0.3.0,<0.4.0)",
    "llama-index-tools-weather (>=0.3.0,<0.4.0)",
    "openai (>=1.60.1,<2.0.0)",
    "tavily-python (>=0.5.0,<0.6.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyowm (>=3.3.0,<4.0.0)",
    "gql (>=3.5.0,<4.0.0)",
    "llama-index-llms-replicate (>=0.4.0,<0.5.0)",
    "replicate (>=1.0.4,<2.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
]

packages = [{ include = "aeris_llamaindex" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
mypy = "^1.14.1"
types-pytz = "^2024.2.0.20241221"
